Reimplemented input handling within the not_implemented screen.
