+++ NeuromancerWin64/NeuromancerWin64.vcxproj
@@ -129,6 +129,7 @@
     <ClCompile Include="level_scene.c" />
     <ClCompile Include="main.c" />
     <ClCompile Include="main_menu_scene.c" />
+    <ClCompile Include="neuro_menu_control.c" />
     <ClCompile Include="neuro_window_control.c" />
     <ClCompile Include="not_implemented_scene.cpp" />
     <ClCompile Include="resource_manager.c" />
@@ -143,6 +144,7 @@
     <ClInclude Include="drawing_control.h" />
     <ClInclude Include="globals.h" />
     <ClInclude Include="inventory_control.h" />
+    <ClInclude Include="neuro_menu_control.h" />
     <ClInclude Include="neuro_window_control.h" />
     <ClInclude Include="resource_manager.h" />
     <ClInclude Include="scene_control.h" />
