+++ NeuromancerWin64/not_implemented_scene.cpp
@@ -1,15 +1,34 @@
 #include "globals.h"
 #include "scene_control.h"
 #include "resource_manager.h"
+#include "neuro_menu_control.h"
 #include "drawing_control.h"
 #include <neuro_routines.h>
 #include <assert.h>
 #include <string.h>
 #include <stdlib.h>
 
+typedef enum not_implemented_state_t {
+	NIS_INITIAL = 0,
+	NIS_TO_MAIN_MENU
+} not_implemented_state_t;
+
+static not_implemented_state_t g_state = NIS_INITIAL;
+
 static uint8_t *g_dialog = NULL;
 static neuro_menu_t g_not_implemented_dialog;
-static int16_t g_selected_dialog_item = -1;
+
+void not_implemented_menu_handle_button_press(int *state, neuro_button_t *button)
+{
+	switch (button->code) {
+	case 0:
+		*state = NIS_TO_MAIN_MENU;
+		break;
+
+	default:
+		break;
+	}
+}
 
 static void init()
 {
@@ -19,77 +38,26 @@ static void init()
 	drawing_control_add_sprite_to_chain(SCI_BACKGRND, 0, 0, g_background, 1);
 
 	assert(g_dialog = calloc(8192, 1));
-	build_menu_dialog_frame(&g_not_implemented_dialog, 64, 100, 192, 40, 6, g_dialog);
-	build_menu_dialog_text(&g_not_implemented_dialog, "Not implemented yet :(", 0, 0);
-	build_menu_dialog_text(&g_not_implemented_dialog, "Okay", 72, 16);
-	build_menu_dialog_item(&g_not_implemented_dialog, 72, 16, 32, 0, 'k');
+	build_neuro_menu_frame(&g_not_implemented_dialog, 64, 100, 192, 40, 6, g_dialog);
+	build_neuro_menu_text(&g_not_implemented_dialog, "Not implemented yet :(", 0, 0);
+	build_neuro_menu_text(&g_not_implemented_dialog, "Okay", 72, 16);
+	build_neuro_menu_item(&g_not_implemented_dialog, 72, 16, 32, 0, 'k');
 	drawing_control_add_sprite_to_chain(SCI_DIALOG, 64, 100, g_dialog, 1);
 }
 
-static neuro_scene_id_t on_menu_dialog_item(char item)
+static void handle_input(sfEvent *event)
 {
-	neuro_scene_id_t scene = NSID_NOT_IMPLEMENTED;
-
-	switch (item)
-	{
-	case 'k':
-		scene = NSID_MAIN_MENU;
-		break;
-
-	default:
-		break;
-	}
-
-	return scene;
+	neuro_menu_handle_input(NMID_MAIN_MENU, &g_not_implemented_dialog, (int*)&g_state, event);
 }
 
 static neuro_scene_id_t update(sfEvent *event)
 {
 	neuro_scene_id_t scene = NSID_NOT_IMPLEMENTED;
-
 	update_cursor();
 
-	/* mouse input */
-	if (sfMouse_isButtonPressed(sfMouseLeft))
-	{
-		int selected = 0;
-
-		for (uint16_t i = 0; i < g_not_implemented_dialog.items_count; i++)
-		{
-			if (cursor_menu_dialog_item_hit_test(i, &g_not_implemented_dialog))
-			{
-				if (g_selected_dialog_item == -1 || g_selected_dialog_item == i)
-				{
-					select_menu_dialog_item(&g_not_implemented_dialog,
-						&g_not_implemented_dialog.items[i], 1);
-					g_selected_dialog_item = i;
-					selected = 1;
-					break;
-				}
-			}
-		}
-		if (!selected)
-		{
-			unselect_menu_dialog_items(&g_not_implemented_dialog);
-		}
-	}
-	else if (event->mouseButton.type == sfEvtMouseButtonReleased)
-	{
-		int selected = g_selected_dialog_item;
-		unselect_menu_dialog_items(&g_not_implemented_dialog);
-		g_selected_dialog_item = -1;
-
-		for (uint16_t i = 0; i < g_not_implemented_dialog.items_count; i++)
-		{
-			if (cursor_menu_dialog_item_hit_test(i, &g_not_implemented_dialog))
-			{
-				if (selected == i)
-				{
-					scene = on_menu_dialog_item(g_not_implemented_dialog.items[i].label);
-					break;
-				}
-			}
-		}
+	switch (g_state) {
+	case NIS_TO_MAIN_MENU:
+		return NSID_MAIN_MENU;
 	}
 
 	return scene;
@@ -106,7 +74,9 @@ void setup_not_implemented_scene()
 {
 	g_scene.id = NSID_NOT_IMPLEMENTED;
 	g_scene.init = init;
-	g_scene.handle_input = NULL;
+	g_scene.handle_input = handle_input;
 	g_scene.update = update;
 	g_scene.deinit = deinit;
+
+	g_state = NIS_INITIAL;
 }
